Moved semantic scholar module
